MINOR: move some client methods to new ClientUtils (#8328)
* move to new ClientUtils

* checkstyle

* fix KafkaStreamsTest

* checkstyle again
